+++ zapisy/apps/offer/proposal/views.py
@@ -192,7 +192,7 @@ def proposal_edit(request, slug=None):
     extrafields = 8
     if request.method == "POST" or (syllabus is not None and len(syllabus.studentwork_set.all())>0):
         extrafields = 1
-    StudentWorkFormset = inlineformset_factory(Syllabus, StudentWork,extra=extrafields)
+    StudentWorkFormset = inlineformset_factory(Syllabus, StudentWork, extra=extrafields, fields='__all__')
     if extrafields == 1:
         student_work_formset = StudentWorkFormset(request.POST or None, instance = syllabus)
     else: