+++ zapisy/apps/users/api_urls.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
-from django.conf.urls import *
+from django.conf.urls import patterns, url
+
 from apps.users.api import StudentList
 
-urlpatterns = patterns('',
-                       url('^student/$', StudentList.as_view())
-)
+urlpatterns = patterns('', url('^student/$', StudentList.as_view()))