+++ zapisy/apps/utils/templatetags/settings_value.py
@@ -3,6 +3,8 @@ Allows templates to access constants
 defined in settings.py. Credit goes
 to Berislav Lopac@stackoverflow
 https://stackoverflow.com/a/7716141
+We've changed it to a filter because Django's
+parser doesn't handle template tags well
 """
 
 from django import template
@@ -10,6 +12,6 @@ from django.conf import settings
 
 register = template.Library()
 
-@register.simple_tag
+@register.filter
 def settings_value(name):
     return getattr(settings, name, "")